Pull out query helpers into own module
