+++ src/lib.rs
@@ -5,6 +5,7 @@ pub mod imports;
 pub mod errors;
 pub mod memory;
 pub mod mock;
+pub mod query;
 pub mod serde;
 pub mod storage;
 pub mod types;
